+++ components/popper/index.js
@@ -11,7 +11,6 @@ const AnimationClassName = 'hi-popper_transition'
 /**
  * @param {Function} onClickOutside 点击该元素外的回调方法
  * @param {Function} setOverlayContainer 获取overLay的挂载父级元素 允许用户指定；如果不指定  popper自己处理
- * @param {Boolean} preventOverflow 防止溢出
  */
 const Popper = (props) => {
   const { show, attachEle, setOverlayContainer } = props