+++ components/select-tree/SelectTreeHook.js
@@ -44,8 +44,12 @@ const SelectTree = ({
   overlayClassName,
   theme,
   localeDatas,
+  placeholder: propsPlaceholder,
+  style,
+  optionWidth,
   placement = 'top-bottom-start'
 }) => {
+  const placeholder = propsPlaceholder || localeDatas.selectTree.placeholder
   const selectedItemsRef = useRef()
   const inputRef = useRef()
   // select 中显示的数量
@@ -199,6 +203,7 @@ const SelectTree = ({
       checked: [],
       semiChecked: []
     })
+    onChange && onChange()
   }, [])
 
   /**
@@ -220,12 +225,15 @@ const SelectTree = ({
     return HiRequest({
       ..._dataSource
     }).then((res) => {
-      const nArr = res.data.map((n) => {
-        return {
-          ...n,
-          pId: id
-        }
-      })
+      const { data = [] } = res
+      const nArr =
+        data &&
+        data.map((n) => {
+          return {
+            ...n,
+            pId: id
+          }
+        })
       return nArr
     })
   }, [])
@@ -327,7 +335,7 @@ const SelectTree = ({
   }
   const searchable = searchMode === 'filter' || searchMode === 'highlight'
   return (
-    <div className={`theme__${theme}`}>
+    <div className={`theme__${theme}`} style={style}>
       <Trigger
         inputRef={inputRef}
         selectedItemsRef={selectedItemsRef}
@@ -336,6 +344,7 @@ const SelectTree = ({
         selectedItems={selectedItems}
         clearable={clearable}
         show={show}
+        placeholder={placeholder}
         checkedEvents={checkedEvents}
         onTrigger={onTrigger}
         onClear={handleClear}
@@ -344,8 +353,8 @@ const SelectTree = ({
         <Popper
           show={show}
           attachEle={inputRef.current}
-          width={false}
           topGap={5}
+          width={optionWidth}
           placement={placement}
           overlayClassName={overlayClassName}
           className={`hi-selecttree__popper ${data.length === 0 && dataSource ? 'hi-selecttree__popper--loading' : ''}`}
@@ -433,7 +442,7 @@ SelectTree.defaultProps = {
   onExpand: () => {},
   checkable: false,
   defaultLoadData: true,
-  showCheckedMode: 'PARENT',
+  showCheckedMode: 'ALL',
   defaultExpandAll: false,
   defaultExpandIds: [],
   expandIds: [],